refactor: move `admin graph` command to `nightscoutadmin graph`
