+++ bot/src/main/java/com/dongtronic/diabot/platforms/discord/commands/admin/AdminCommand.kt
@@ -22,7 +22,7 @@ class AdminCommand(category: Command.Category) : DiscordCommand(category, null) 
                 AdminRewardsCommand(category, this),
                 AdminChannelsCommand(category, this),
                 AdminAnnounceCommand(category, this),
-                AdminGraphCommand(category, this))
+        )
     }
 
     override fun execute(event: CommandEvent) {
